Some compilation errors and all x64 warnings fixed

+++ cegui/include/CEGUI/GeometryBuffer.h
@@ -542,7 +542,7 @@ protected:
     //! The activation number of the RenderTarget that this GeometryBuffer's matrix was last updated for
     mutable unsigned int    d_lastRenderTargetActivationCount = 0;
     //! The vertex count which is determined based on the used vertex layout
-    unsigned int                d_vertexCount = 0;
+    size_t                d_vertexCount = 0;
     //! The amount of vertices that need to be rendered after rendering to the stencil buffer.
     unsigned int    d_postStencilVertexCount = 0;
     //! The BlendMode to use when rendering this GeometryBuffer.
